+++ src/Resources/contao/controllers/FrontendIndex.php
@@ -32,9 +32,11 @@ class FrontendIndex extends Frontend
 		$this->import('FrontendUser', 'User');
 		parent::__construct();
 
+		$objTokenChecker = \System::getContainer()->get('contao.security.token_checker');
+
 		// Check whether a user is logged in
-		\define('BE_USER_LOGGED_IN', $this->getLoginStatus('BE_USER_AUTH'));
-		\define('FE_USER_LOGGED_IN', $this->getLoginStatus('FE_USER_AUTH'));
+		\define('BE_USER_LOGGED_IN', $objTokenChecker->hasBackendUser() && $objTokenChecker->isPreviewMode());
+		\define('FE_USER_LOGGED_IN', $objTokenChecker->hasFrontendUser());
 	}
 
 	/**
@@ -128,6 +130,12 @@ class FrontendIndex extends Frontend
 				$arrLangs = $arrPages['*'] ?: array(); // empty domain
 			}
 
+			// Throw an exception if there are no matches (see #1522)
+			if (empty($arrLangs))
+			{
+				throw new PageNotFoundException('Page not found: ' . \Environment::get('uri'));
+			}
+
 			// Use the first result (see #4872)
 			if (!\Config::get('addLanguageToUrl'))
 			{
@@ -169,16 +177,10 @@ class FrontendIndex extends Frontend
 		// If the page has an alias, it can no longer be called via ID (see #7661)
 		if ($objPage->alias != '')
 		{
-			if (\Config::get('addLanguageToUrl'))
-			{
-				$regex = '#^[a-z]{2}(-[A-Z]{2})?/' . $objPage->id . '[$/.]#';
-			}
-			else
-			{
-				$regex = '#^' . $objPage->id . '[$/.]#';
-			}
+			$language = \Config::get('addLanguageToUrl') ? '[a-z]{2}(-[A-Z]{2})?/' : '';
+			$suffix = \Config::get('urlSuffix') ? preg_quote(\Config::get('urlSuffix'), '#') : '$';
 
-			if (preg_match($regex, \Environment::get('relativeRequest')))
+			if (preg_match('#^' . $language . $objPage->id . '(' . $suffix . '|/)#', \Environment::get('relativeRequest')))
 			{
 				throw new PageNotFoundException('Page not found: ' . \Environment::get('uri'));
 			}