+++ src/Resources/contao/elements/ContentGallery.php
@@ -106,7 +106,6 @@ class ContentGallery extends ContentElement
 					'uuid'       => $objFiles->uuid,
 					'name'       => $objFile->basename,
 					'singleSRC'  => $objFiles->path,
-					'title'      => \StringUtil::specialchars($objFile->basename),
 					'filesModel' => $objFiles->current()
 				);
 
@@ -145,7 +144,6 @@ class ContentGallery extends ContentElement
 						'uuid'       => $objSubfiles->uuid,
 						'name'       => $objFile->basename,
 						'singleSRC'  => $objSubfiles->path,
-						'title'      => \StringUtil::specialchars($objFile->basename),
 						'filesModel' => $objSubfiles->current()
 					);
 
@@ -253,8 +251,7 @@ class ContentGallery extends ContentElement
 
 		$rowcount = 0;
 		$colwidth = floor(100/$this->perRow);
-		$intMaxWidth = (TL_MODE == 'BE') ? floor((640 / $this->perRow)) : floor((\Config::get('maxImageWidth') / $this->perRow));
-		$strLightboxId = 'lightbox[lb' . $this->id . ']';
+		$strLightboxId = 'lb' . $this->id;
 		$body = array();
 
 		// Rows
@@ -305,7 +302,7 @@ class ContentGallery extends ContentElement
 					$images[($i+$j)]['imagemargin'] = $this->imagemargin;
 					$images[($i+$j)]['fullsize'] = $this->fullsize;
 
-					$this->addImageToTemplate($objCell, $images[($i+$j)], $intMaxWidth, $strLightboxId, $images[($i+$j)]['filesModel']);
+					$this->addImageToTemplate($objCell, $images[($i+$j)], null, $strLightboxId, $images[($i+$j)]['filesModel']);
 
 					// Add column width and class
 					$objCell->colWidth = $colwidth . '%';