+++ src/Resources/contao/library/Contao/Controller.php
@@ -16,6 +16,7 @@ use Contao\CoreBundle\Exception\AjaxRedirectResponseException;
 use Contao\CoreBundle\Exception\PageNotFoundException;
 use Contao\CoreBundle\Exception\RedirectResponseException;
 use Contao\Database\Result;
+use Contao\Image\PictureConfigurationInterface;
 use League\Uri\Components\Query;
 use Symfony\Component\Finder\Finder;
 use Symfony\Component\Finder\Glob;
@@ -1442,16 +1443,19 @@ abstract class Controller extends System
 		{
 			$size = array(0, 0, (int) $size);
 		}
-		elseif (!\is_array($size))
+		elseif (!$size instanceof PictureConfigurationInterface)
 		{
-			$size = array();
-		}
+			if (!\is_array($size))
+			{
+				$size = array();
+			}
 
-		$size += array(0, 0, 'crop');
+			$size += array(0, 0, 'crop');
+		}
 
 		if ($intMaxWidth === null)
 		{
-			$intMaxWidth = (TL_MODE == 'BE') ? 320 : \Config::get('maxImageWidth');
+			$intMaxWidth = \Config::get('maxImageWidth');
 		}
 
 		$arrMargin = (TL_MODE == 'BE') ? array() : \StringUtil::deserialize($arrItem['imagemargin']);
@@ -1463,6 +1467,8 @@ abstract class Controller extends System
 		// Adjust the image size
 		if ($intMaxWidth > 0)
 		{
+			@trigger_error('Using a maximum front end width has been deprecated and will no longer work in Contao 5.0. Remove the "maxImageWidth" configuration and use responsive images instead.', E_USER_DEPRECATED);
+
 			// Subtract the margins before deciding whether to resize (see #6018)
 			if (\is_array($arrMargin) && $arrMargin['unit'] == 'px')
 			{
@@ -1480,7 +1486,7 @@ abstract class Controller extends System
 				}
 			}
 
-			if ($size[0] > $intMaxWidth || (!$size[0] && !$size[1] && (!$imgSize[0] || $imgSize[0] > $intMaxWidth)))
+			if (\is_array($size) && ($size[0] > $intMaxWidth || (!$size[0] && !$size[1] && (!$imgSize[0] || $imgSize[0] > $intMaxWidth))))
 			{
 				// See #2268 (thanks to Thyon)
 				$ratio = ($size[0] && $size[1]) ? $size[1] / $size[0] : (($imgSize[0] && $imgSize[1]) ? $imgSize[1] / $imgSize[0] : 0);
@@ -1490,17 +1496,10 @@ abstract class Controller extends System
 			}
 		}
 
-		// Disable responsive images in the back end (see #7875)
-		if (TL_MODE == 'BE')
-		{
-			unset($size[2]);
-		}
-
 		try
 		{
 			$container = \System::getContainer();
 			$staticUrl = $container->get('contao.assets.files_context')->getStaticUrl();
-			$src = $container->get('contao.image.image_factory')->create(TL_ROOT . '/' . $arrItem['singleSRC'], $size)->getUrl(TL_ROOT);
 			$picture = $container->get('contao.image.picture_factory')->create(TL_ROOT . '/' . $arrItem['singleSRC'], $size);
 
 			$picture = array
@@ -1508,6 +1507,8 @@ abstract class Controller extends System
 				'img' => $picture->getImg(TL_ROOT, $staticUrl),
 				'sources' => $picture->getSources(TL_ROOT, $staticUrl)
 			);
+
+			$src = $picture['img']['src'];
 
 			if ($src !== $arrItem['singleSRC'])
 			{
@@ -1604,7 +1605,7 @@ abstract class Controller extends System
 		// Provide an ID for single lightbox images in HTML5 (see #3742)
 		if ($strLightboxId === null && $arrItem['fullsize'])
 		{
-			$strLightboxId = 'lightbox[' . substr(md5($objTemplate->getName() . '_' . $arrItem['id']), 0, 6) . ']';
+			$strLightboxId = substr(md5($objTemplate->getName() . '_' . $arrItem['id']), 0, 6);
 		}
 
 		// Float image
@@ -1633,7 +1634,7 @@ abstract class Controller extends System
 						$objTemplate->$strHrefKey = static::addFilesUrlTo(\System::urlEncode($arrItem['imageUrl']));
 					}
 
-					$objTemplate->attributes = ' data-lightbox="' . substr($strLightboxId, 9, -1) . '"';
+					$objTemplate->attributes = ' data-lightbox="' . $strLightboxId . '"';
 				}
 				else
 				{
@@ -1646,7 +1647,7 @@ abstract class Controller extends System
 		elseif ($arrItem['fullsize'] && TL_MODE == 'FE')
 		{
 			$objTemplate->$strHrefKey = static::addFilesUrlTo(\System::urlEncode($arrItem['singleSRC']));
-			$objTemplate->attributes = ' data-lightbox="' . substr($strLightboxId, 9, -1) . '"';
+			$objTemplate->attributes = ' data-lightbox="' . $strLightboxId . '"';
 		}
 
 		// Add the meta data to the template