+++ src/Resources/contao/models/ContentModel.php
@@ -69,7 +69,6 @@ namespace Contao;
  * @property string  $youtube
  * @property string  $posterSRC
  * @property string  $playerSize
- * @property boolean $autoplay
  * @property integer $sliderDelay
  * @property integer $sliderSpeed
  * @property integer $sliderStartSlide
@@ -94,7 +93,7 @@ namespace Contao;
  * @property boolean $com_requireLogin
  * @property string  $com_template
  * @property string  $typePrefix
- * @property string  $classes
+ * @property array   $classes
  * @property integer $origId
  *
  * @method static ContentModel|null findById($id, array $opt=array())